test(suite): co-locate hook tests
